+++ src/games/next-station-london/constants.ts
@@ -1,0 +1,5 @@
+export const BOARD_SIZE = 10;
+export const OUTPUT_SIZE = 1000;
+export const GRID_SIZE = OUTPUT_SIZE / BOARD_SIZE;
+export const NODE_SIZE = GRID_SIZE / 2;
+export const SHAPE_SIZE = NODE_SIZE - 25;
